Suppress error on resetUserStat()
Suppose we call resetUserStat() two times consequtively.  First time,
the user stat will be reset, and that will be commit()ted.  However,
second time, there will be no change, and git will complain that there
is nothing to commit, but it makes no sense to complain this to user.
Thus we will just suppress it in resetUserStat().